moved resume host zone test
